Using Yacc and Lex and C programming
In this project we are going to analyze the DNA sequence pattern entered by user and check it whether the give DNA sequence is valid or not. For this purpose the DNA pattern entered should belong to the specific language accepted by the analyzer which will be defined by us. Also we are going to give option to user to enter more than one DNA sequence and then we will check there validity and also whether the hybridization between the two DNA entered is possible or not. Hence it would form a lexical analyzer along with the computation from the given strings of the language. So with this we can easily check whether the DNA strands are properly entered with respect to the language and whether we can hybridize them to form newly hybridized DNAs by checking some similarity conditions for each pair.